Simplified cycle detection.
Cycles are only created when a path (A, B) is created in a graph that already
has a path (B, A).  The self-check is unnecessary and leads to semantic
problems, such as:

  link.save  # => true
  link.save  # => false

It is still important to prevent two copies of the same edge from existing in
the database.  Unfortunately, we need a non-compound primary key; ActiveRecord
requires this.

The composite_primary_keys gem relieves that requirement, but I'm hesitant to
drag that in without knowing about the impact of other ways to enforce
uniqueness, such as indices.