Re-create generated Node associations for derived classes.
The Node associations (links_as_child, links_as_ancestor, etc.) contain a
condition on ancestor or descendant type.  This condition is derived from the
name of the model class acting as a dag node.

Therefore, for derived model classes, this condition can be properly created
only in the derived class.

This changeset modifies Dagnabit::Node::Associations to re-install associations
in a subclass whenever a base model class is subclassed.