Validate ancestor and descendant when saving links.
As a bonus, this makes Link#connect return false when links cannot be made
between two nodes due to validation failure or whatnot.  Thus, instead of writing

  ok = node2.save
  Link.connect(node1, node2) if ok

it is now possible to just write

  Link.connect(node1, node2)

node1 and node2 will be saved (if necessary).  If node1 or node2 fail
validations, the link will not be saved.

However, this change prevents code like this from working:

  node1.links_as_parent.build(:descendant => node2)

or

  node2.links_as_child.build(:ancestor => node1)

The reason why is that the link between node1 and node2 has a null ancestor
(or, in the second example, a null descendant).  The ancestor/descendant is
actually present, but is not recognized as such at validation time.

ActiveRecord 3.0 has a fix for this in the form of the inverse_of patch, but it
appears to not play nicely with the validates_associated changes.  (In
particular, you end up with circular dependencies between the link and node.)