Revert "Destroy links to parents and children on node deletion."
This reverts commit 9c63f8e47187f9a749a1e311f1cbef69f8fed6ff.

This code isn't ready yet: it has at least one showstopper error.  The error
manifests in graphs that are their own transitive closure, e.g.

  a1 -> a2 -> a3
  |___________|

If a2 were to be deleted, then one would expect the following:

  a1          a3
  |___________|

Instead, what you get is

  a1          a3

which is obviously not right.  The deletion algorithm appears to be
overzealous, but it's not yet entirely clear what part of it is the offender.