Fix transitive closure update error caused by an incorrect "trusty" edge finder.
This commit also re-enables destruction of links to children and parents
on node destruction.

When we recompute the graph's transitive closure after a link (A, B) is
destroyed, one of the ways in which we find "trusty" edges is to include
all remaining edges in the graph that are not (A, B).

The previous algorithm found edges using the expression

  (ancestor <> A and descendant <> B)

which is incorrect, as it would exclude any edges for which one _but
not_ the other was true.

The correct expression is

  not (ancestor = A and descendant = B)

which is what this changeset includes.