Load ActiveRecord before activating dagnabit.
This avoids a load order anomaly that occurs particularly often when
using dagnabit with bundler, in which dagnabit's additions to
ActiveRecord are overwritten.

No tests because I (unfortunately) don't yet fully understand the
problem; I only know it happens and that ensuring that ActiveRecord is
loaded fixes my issue.  When I understand the problem better, that's
when we can do a test.