Dagnabit::Connectivity -> Dagnabit::Vertex::Connectivity.
I'll be introducing a method to find edges connecting vertices (it'd
be convenient to have for a project that uses dagnabit), and separating
out queries that find edges vs. queries that find vertices is something
that I find useful.  (If nothing else, it provides a nice structure for
writing examples.)