Propagate edge table settings to subclasses.
The method in this commit is kind of gross.  It might be better to use
class_inheritable_accessor or some such from ActiveSupport.
Unfortunately, I can't find a way of doing that preserves my goal of
keeping

    extend Dagnabit::Vertex::Connectivity

(or, alternatively, include) the only thing needed to activate dagnabit
on a model object.  (The included/extended callbacks don't offer any
advantage over the inherited callback-based method that I'm using.)