Move all dagnabit setup for specs into test models.
This is the easiest way I've found that satisfies
ActiveRecord::Associations' assumption that all ActiveRecord models are
bound to a constant.  I was running into too many spec failures related
to me violating that assumption before making this change, and every
attempt I made to fix it whilst preserving the old ways just made the
specs less readable.

It does mean that we can't specify which modules are prerequisites for
other modules anymore, and I'm disappointed to lose that, but for now I
think it's an acceptable tradeoff.