Simplified the mutators of Vertex::Settings to one method: Settings#connected_by.
Settings#connected_by accepts the name of an edge model.  #edge_model
constantizes it, and #edge_table has been redefined as
`edge_model.table_name`.

This means that an edge model is now required, but I don't think that's
a big deal.  (It definitely isn't a big deal for the one user that
dagnabit 3 has, i.e. me.)