First cut at Dagnabit::Vertex::Associations.
I added the in_edges and out_edges associations because I needed a
convenient way to have edges destroyed on vertex destruction.

The addition of foreign keys to the test schema also plays into this
need.  Applications that use dagnabit may have foreign keys from their
edge tables to their vertex tables, so we should make sure that
vertex destruction plays nice with that.

The intent is not to test ActiveRecord's associations: those (usually)
work.  The intent is to make sure that we didn't forget to pass the
right options to said associations.