Hack around a foreign key derivation error in composite_primary_keys.
First off: Yes, I know this is the kind of thing that should be fixed in
CPK.  However, fixing this in CPK would require me to

  1. understand its test suite well enough to know how to write a test
     that demonstrates the failure
  2. write the test
  3. fix the problem
  4. submit a patch
  5. wait for upstream to accept it
  6. possibly go to 2, depending on the outcome of 5

Items 1 and 5 can take quite a bit of time, and I don't have that time
to spare.  Maybe I will later.

Anyway, a description of the problem:

ActiveRecord derives foreign key names for belongs_to associations
using, in order of precedence,

  1. the value of the :foreign_key option
  2. the name of the association + "_id"

CPK 2.3.5.1 (actually, CPK <= 3.1.6) do this:

  1. the value of the :foreign_key option
  2. the foreign key of the class name of the belongs_to reflection

Thus, given an application using CPK and a graph setup in which the
vertices are represented by a Vertex model and edges by an Edge model,
the parent and child associations on Edge will both reference a
nonexistent "vertex_id" foreign key field on Edge.  This causes pretty
much everything related to setting vertices on edges to fail.

This commit works around the problem by explicitly specifying foreign
key names, which shortcuts the foreign key derivation codepaths in both
AR and CPK.